requirements.txt
Flask==3.1.1
paho_mqtt==2.1.0
python-dotenv==1.1.1
